Opdatere tekst, placering og mapper/navne
Opdatere tekst for siderne:
 - Hjem
 - Sidebar
Opdater placering for:
 - Forberedelse
 - Sidebar
Opdateret navne for:
 - UgeplanSidebar -> Sidebar
