+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
 import App from './App.tsx';
-import UgeplanSidebar from './components/UgeplanSidebar.tsx';
+import Sidebar from './components/Sidebar.tsx';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
@@ -10,7 +10,7 @@ createRoot(document.getElementById('root')!).render(
       <div className="container">
         <div className="row">
           <div className="col-3">
-            <UgeplanSidebar/>
+            <Sidebar/>
           </div>
           <div className="col-9" style={{marginTop: 15}}>
             <App /> {/* App handles the routes and layout */}
