--- src/pages/Ugeplan.tsx
@@ -1,27 +1,0 @@
-function Ugeplan() {
-  
-    return (
-      <>
-        <div className="container">
-          <div className="row">
-            <div className="col-9">
-
-              <p>Waddup</p>
-              <div class="list-group">
-                <button type="button" className="list-group-item list-group-item-action active" aria-current="true">
-                  The current button
-                </button>
-                <button type="button" className="list-group-item list-group-item-action">A second button item</button>
-                <button type="button" className="list-group-item list-group-item-action">A third button item</button>
-                <button type="button" className="list-group-item list-group-item-action">A fourth button item</button>
-                <button type="button" className="list-group-item list-group-item-action" disabled>A disabled button item</button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </>
-    )
-  }
-  
-  export default Ugeplan
-
